requirements.txt
llama-index==0.5.17
botbuilder-integration-aiohttp @ git+https://github.com/LGDoor/botbuilder-python.git@main#subdirectory=libraries/botbuilder-integration-aiohttp
langchain==0.0.144
nltk==3.8.1
